Revert "SOLR-17302: Convert /cluster filestore APIs to JAX-RS (#2470)"
This reverts commit 8c5ae1bb745869ebd6fab5911b3dd0fd75ed9106.
