+++ solr/core/src/java/org/apache/solr/filestore/DistribFileStore.java
@@ -216,9 +216,8 @@ public class DistribFileStore implements FileStore {
     }
 
     boolean fetchFromAnyNode() {
-      ArrayList<String> nodesToAttemptFetchFrom =
-          FileStoreUtils.fetchAndShuffleRemoteLiveNodes(coreContainer);
-      for (String liveNode : nodesToAttemptFetchFrom) {
+      ArrayList<String> l = coreContainer.getFileStoreAPI().shuffledNodes();
+      for (String liveNode : l) {
         try {
           String baseUrl =
               coreContainer.getZkController().getZkStateReader().getBaseUrlV2ForNodeName(liveNode);
@@ -243,6 +242,12 @@ public class DistribFileStore implements FileStore {
       }
 
       return false;
+    }
+
+    String getSimpleName() {
+      int idx = path.lastIndexOf('/');
+      if (idx == -1) return path;
+      return path.substring(idx + 1);
     }
 
     public Path realPath() {
@@ -301,13 +306,6 @@ public class DistribFileStore implements FileStore {
           ew.put("timestamp", getTimeStamp());
           if (metaData != null) metaData.writeMap(ew);
         }
-
-        @Override
-        public String getSimpleName() {
-          int idx = path.lastIndexOf('/');
-          if (idx == -1) return path;
-          return path.substring(idx + 1);
-        }
       };
     }
 
@@ -355,7 +353,7 @@ public class DistribFileStore implements FileStore {
     }
     tmpFiles.put(info.path, info);
 
-    List<String> nodes = FileStoreUtils.fetchAndShuffleRemoteLiveNodes(coreContainer);
+    List<String> nodes = coreContainer.getFileStoreAPI().shuffledNodes();
     int i = 0;
     int FETCHFROM_SRC = 50;
     String myNodeName = coreContainer.getZkController().getNodeName();
@@ -498,7 +496,7 @@ public class DistribFileStore implements FileStore {
   @Override
   public void delete(String path) {
     deleteLocal(path);
-    List<String> nodes = FileStoreUtils.fetchAndShuffleRemoteLiveNodes(coreContainer);
+    List<String> nodes = coreContainer.getFileStoreAPI().shuffledNodes();
     HttpClient client = coreContainer.getUpdateShardHandler().getDefaultHttpClient();
     for (String node : nodes) {
       String baseUrl =
@@ -585,7 +583,7 @@ public class DistribFileStore implements FileStore {
   }
 
   public static synchronized Path getFileStoreDirPath(Path solrHome) {
-    var path = solrHome.resolve(ClusterFileStore.FILESTORE_DIRECTORY);
+    var path = solrHome.resolve(FileStoreAPI.FILESTORE_DIRECTORY);
     if (!Files.exists(path)) {
       try {
         Files.createDirectories(path);
@@ -634,7 +632,7 @@ public class DistribFileStore implements FileStore {
   // reads local keys file
   private static Map<String, byte[]> _getKeys(Path solrHome) throws IOException {
     Map<String, byte[]> result = new HashMap<>();
-    Path keysDir = _getRealPath(ClusterFileStore.KEYS_DIR, solrHome);
+    Path keysDir = _getRealPath(FileStoreAPI.KEYS_DIR, solrHome);
 
     File[] keyFiles = keysDir.toFile().listFiles();
     if (keyFiles == null) return result;
