+++ solr/core/src/java/org/apache/solr/filestore/FileStore.java
@@ -111,8 +111,6 @@ public interface FileStore {
 
   interface FileDetails extends MapWriter {
 
-    String getSimpleName();
-
     MetaData getMetaData();
 
     Date getTimeStamp();
