+++ solr/core/src/java/org/apache/solr/pkg/PackageAPI.java
@@ -45,7 +45,7 @@ import org.apache.solr.common.util.EnvUtils;
 import org.apache.solr.common.util.ReflectMapWriter;
 import org.apache.solr.common.util.Utils;
 import org.apache.solr.core.CoreContainer;
-import org.apache.solr.filestore.FileStoreUtils;
+import org.apache.solr.filestore.FileStoreAPI;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
 import org.apache.solr.util.SolrJacksonAnnotationInspector;
@@ -73,6 +73,10 @@ public class PackageAPI {
   public final Read readAPI = new Read();
 
   public PackageAPI(CoreContainer coreContainer, SolrPackageLoader loader) {
+    if (coreContainer.getFileStoreAPI() == null) {
+      throw new IllegalStateException("Must successfully load FileStoreAPI first");
+    }
+
     this.coreContainer = coreContainer;
     this.packageLoader = loader;
     pkgs = new Packages();
@@ -250,10 +254,10 @@ public class PackageAPI {
       }
       // first refresh my own
       packageLoader.notifyListeners(p);
-      for (String liveNode : FileStoreUtils.fetchAndShuffleRemoteLiveNodes(coreContainer)) {
+      for (String s : coreContainer.getFileStoreAPI().shuffledNodes()) {
         Utils.executeGET(
             coreContainer.getUpdateShardHandler().getDefaultHttpClient(),
-            coreContainer.getZkController().zkStateReader.getBaseUrlV2ForNodeName(liveNode)
+            coreContainer.getZkController().zkStateReader.getBaseUrlV2ForNodeName(s)
                 + "/cluster/package?wt=javabin&omitHeader=true&refreshPackage="
                 + p,
             Utils.JAVABINCONSUMER);
@@ -268,8 +272,8 @@ public class PackageAPI {
         payload.addError("No files specified");
         return;
       }
-      FileStoreUtils.validateFiles(
-          coreContainer.getFileStore(), add.files, true, s -> payload.addError(s));
+      FileStoreAPI fileStoreAPI = coreContainer.getFileStoreAPI();
+      fileStoreAPI.validateFiles(add.files, true, s -> payload.addError(s));
       if (payload.hasError()) return;
       Packages[] finalState = new Packages[1];
       try {
@@ -418,10 +422,10 @@ public class PackageAPI {
   }
 
   void notifyAllNodesToSync(int expected) {
-    for (String liveNode : FileStoreUtils.fetchAndShuffleRemoteLiveNodes(coreContainer)) {
+    for (String s : coreContainer.getFileStoreAPI().shuffledNodes()) {
       Utils.executeGET(
           coreContainer.getUpdateShardHandler().getDefaultHttpClient(),
-          coreContainer.getZkController().zkStateReader.getBaseUrlV2ForNodeName(liveNode)
+          coreContainer.getZkController().zkStateReader.getBaseUrlV2ForNodeName(s)
               + "/cluster/package?wt=javabin&omitHeader=true&expectedVersion"
               + expected,
           Utils.JAVABINCONSUMER);
